实现了购物车的简单添加和查看
完善了部分结构层次

+++ handler/product.go
@@ -79,7 +79,7 @@ func GetProductDetail(_ context.Context, ctx *app.RequestContext) {
 		return
 	}
 
-	response, err := service.GetProductDetail(productID)
+	response, err := service.GetProductDetail(uint(productID))
 	if err != nil {
 		ctx.JSON(consts.StatusOK, types.ErrorResponse(10109, "获取商品详情失败"))
 		return
